:truck: Rename packages/localization/intl/index.tsx to index.ts
